// Bimaru.Frontend.Angular/angularapp/src/app/bimaru/interfaces/bimaruValue.ts
export enum BimaruValue {
    Undetermined = "UNDETERMINED",
    Water = "WATER",
    ShipContinuedLeft = "SHIP_CONT_LEFT",
    ShipContinuedUp = "SHIP_CONT_UP",
    ShipContinuedRight = "SHIP_CONT_RIGHT",
    ShipContinuedDown = "SHIP_CONT_DOWN",
    ShipMiddle = "SHIP_MIDDLE",
    ShipSingle = "SHIP_SINGLE",
    ShipUndetermined = "SHIP_UNDETERMINED"
}